ncaa_bb_games

Obtains basketball game data by date from the ncaa stats API. Some notes:

When values are missing (such as a team rank, or a bracket round), they are recorded as NA

dates and times are recorded in ET as is returned by the ET

If seeking to group by a team identifier, it is recommended to use slugs

game_id can be used to obtain pbp and box score data for the game

ncaa_bb_games(date = Sys.Date(), type = "men")

Arguments

date

An R Date object. Defaults to Sys.Date()

type

Men's or Women's game schedule (options: 'men', 'women')

Value

a data frame with NCAA MBB game data

Examples

# \donttest{
  ncaa_bb_games(as.Date('2021-12-03'))
#>    game_id          start_time       date bracket_round bracket_id
#> 1  5897924 2021-12-03 22:00:00 2021-12-03          <NA>       <NA>
#> 2  5897932 2021-12-03 23:00:00 2021-12-03          <NA>       <NA>
#> 3  5897923 2021-12-04 00:00:00 2021-12-03          <NA>       <NA>
#> 4  5897927 2021-12-04 00:00:00 2021-12-03          <NA>       <NA>
#> 5  5897928 2021-12-04 00:00:00 2021-12-03          <NA>       <NA>
#> 6  5897930 2021-12-04 00:00:00 2021-12-03          <NA>       <NA>
#> 7  5897934 2021-12-04 00:00:00 2021-12-03          <NA>       <NA>
#> 8  5897935 2021-12-04 00:00:00 2021-12-03          <NA>       <NA>
#> 9  5897936 2021-12-04 00:00:00 2021-12-03          <NA>       <NA>
#> 10 5897937 2021-12-04 00:00:00 2021-12-03          <NA>       <NA>
#> 11 5897938 2021-12-04 00:00:00 2021-12-03          <NA>       <NA>
#> 12 5897940 2021-12-04 00:00:00 2021-12-03          <NA>       <NA>
#> 13 5897943 2021-12-04 00:00:00 2021-12-03          <NA>       <NA>
#> 14 5897945 2021-12-04 00:00:00 2021-12-03          <NA>       <NA>
#> 15 5897947 2021-12-04 00:00:00 2021-12-03          <NA>       <NA>
#> 16 5897925 2021-12-04 01:00:00 2021-12-03          <NA>       <NA>
#> 17 5897931 2021-12-04 01:00:00 2021-12-03          <NA>       <NA>
#> 18 5897939 2021-12-04 01:00:00 2021-12-03          <NA>       <NA>
#> 19 5897941 2021-12-04 01:00:00 2021-12-03          <NA>       <NA>
#> 20 5897942 2021-12-04 01:00:00 2021-12-03          <NA>       <NA>
#> 21 5897949 2021-12-04 01:00:00 2021-12-03          <NA>       <NA>
#> 22 5897929 2021-12-04 02:00:00 2021-12-03          <NA>       <NA>
#> 23 5897950 2021-12-04 02:00:00 2021-12-03          <NA>       <NA>
#> 24 5897926 2021-12-04 03:00:00 2021-12-03          <NA>       <NA>
#> 25 5897944 2021-12-04 03:00:00 2021-12-03          <NA>       <NA>
#> 26 5897946 2021-12-04 03:00:00 2021-12-03          <NA>       <NA>
#> 27 5897948 2021-12-04 03:00:00 2021-12-03          <NA>       <NA>
#> 28 5909927 2021-12-04 03:00:00 2021-12-03          <NA>       <NA>
#>    bracket_region game_network game_status current_game_period
#> 1            <NA>         <NA>       final               FINAL
#> 2            <NA>         <NA>       final               FINAL
#> 3            <NA>         <NA>       final               FINAL
#> 4            <NA>         <NA>       final               FINAL
#> 5            <NA>         <NA>       final               FINAL
#> 6            <NA>         <NA>       final               FINAL
#> 7            <NA>         <NA>       final               FINAL
#> 8            <NA>         <NA>       final               FINAL
#> 9            <NA>         <NA>       final               FINAL
#> 10           <NA>         <NA>       final               FINAL
#> 11           <NA>         <NA>       final               FINAL
#> 12           <NA>         <NA>       final               FINAL
#> 13           <NA>         <NA>       final               FINAL
#> 14           <NA>         <NA>       final               FINAL
#> 15           <NA>         <NA>       final               FINAL
#> 16           <NA>         <NA>       final               FINAL
#> 17           <NA>         <NA>       final               FINAL
#> 18           <NA>         <NA>       final               FINAL
#> 19           <NA>         <NA>       final               FINAL
#> 20           <NA>         <NA>       final               FINAL
#> 21           <NA>         <NA>       final               FINAL
#> 22           <NA>         <NA>       final               FINAL
#> 23           <NA>         <NA>       final               FINAL
#> 24           <NA>         <NA>       final               FINAL
#> 25           <NA>         <NA>       final               FINAL
#> 26           <NA>         <NA>       final               FINAL
#> 27           <NA>         <NA>       final               FINAL
#> 28           <NA>         <NA>       final               FINAL
#>    current_game_clock away_score away_abbrev         away_short
#> 1                0:00         55      CARBIB             Carver
#> 2                0:00         57      N DAME         Notre Dame
#> 3                0:00         88      C OF C Col. of Charleston
#> 4                0:00         65       SC ST South Carolina St.
#> 5                0:00         51      RUTGER            Rutgers
#> 6                0:00         61      BOB JO          Bob Jones
#> 7                0:00         95      KANSAS             Kansas
#> 8                0:00         69      COLGAT            Colgate
#> 9                0:00         69      ST PTR      Saint Peter's
#> 10               0:00         74        FAIR          Fairfield
#> 11               0:00         57      MONMTH           Monmouth
#> 12               0:00         77      MANHAT          Manhattan
#> 13               0:00         54       RIDER              Rider
#> 14               0:00         58       UTRGV              UTRGV
#> 15               0:00         77      COPPIN         Coppin St.
#> 16               0:00         44      BRYANT             Bryant
#> 17               0:00         67         UNO        New Orleans
#> 18               0:00         72       NM ST     New Mexico St.
#> 19               0:00         47      MNMORR       Minn.-Morris
#> 20               0:00         37      WALDRF            Waldorf
#> 21               0:00         56        PITT         Pittsburgh
#> 22               0:00         58       TULSA              Tulsa
#> 23               0:00         70        IOWA               Iowa
#> 24               0:00         59      BNMESA   Benedictine Mesa
#> 25               0:00         90         VMI                VMI
#> 26               0:00         51      NO DAK       North Dakota
#> 27               0:00         74      PEPPER         Pepperdine
#> 28               0:00         54      BETCHS      Bethesda (CA)
#>             away_slug
#> 1        carver-bible
#> 2          notre-dame
#> 3   col-of-charleston
#> 4   south-carolina-st
#> 5             rutgers
#> 6           bob-jones
#> 7              kansas
#> 8             colgate
#> 9           st-peters
#> 10          fairfield
#> 11           monmouth
#> 12          manhattan
#> 13              rider
#> 14              utrgv
#> 15          coppin-st
#> 16             bryant
#> 17        new-orleans
#> 18      new-mexico-st
#> 19        minn-morris
#> 20            waldorf
#> 21         pittsburgh
#> 22              tulsa
#> 23               iowa
#> 24   benedictine-mesa
#> 25                vmi
#> 26       north-dakota
#> 27         pepperdine
#> 28 bethesda-christian
#>                                                     away_full away_rank
#> 1                                                      Carver      <NA>
#> 2                                    University of Notre Dame      <NA>
#> 3                      College of Charleston (South Carolina)      <NA>
#> 4                             South Carolina State University      <NA>
#> 5  Rutgers, The State University of New Jersey, New Brunswick      <NA>
#> 6                                        Bob Jones University      <NA>
#> 7                                        University of Kansas         8
#> 8                                          Colgate University      <NA>
#> 9                                    Saint Peter's University      <NA>
#> 10                                       Fairfield University      <NA>
#> 11                                        Monmouth University      <NA>
#> 12                                          Manhattan College      <NA>
#> 13                                           Rider University      <NA>
#> 14                  The University of Texas Rio Grande Valley      <NA>
#> 15                                    Coppin State University      <NA>
#> 16                                          Bryant University      <NA>
#> 17                                  University of New Orleans      <NA>
#> 18                                New Mexico State University      <NA>
#> 19                            University of Minnesota, Morris      <NA>
#> 20                                                    Waldorf      <NA>
#> 21                                   University of Pittsburgh      <NA>
#> 22                                    The University of Tulsa      <NA>
#> 23                                         University of Iowa      <NA>
#> 24                                           Benedictine Mesa      <NA>
#> 25                                Virginia Military Institute      <NA>
#> 26                                 University of North Dakota      <NA>
#> 27                                      Pepperdine University      <NA>
#> 28                                              Bethesda (CA)      <NA>
#>    away_seed away_conference home_score home_abbrev       home_short
#> 1       <NA>    NON-NCAA ORG         97      NC A&T         N.C. A&T
#> 2       <NA>             ACC         73          BC   Boston College
#> 3       <NA>             CAA         91      FURMAN           Furman
#> 4       <NA>            MEAC         64         USF       South Fla.
#> 5       <NA>         Big Ten         86         ILL         Illinois
#> 6       <NA>    NON-NCAA ORG         90      PRESBY     Presbyterian
#> 7       <NA>          Big 12         75      STJOHN  St. John's (NY)
#> 8       <NA>         Patriot         74      NOEAST     Northeastern
#> 9       <NA>            MAAC         59       QUINN       Quinnipiac
#> 10      <NA>            MAAC         68      CANISI         Canisius
#> 11      <NA>            MAAC         49      NIAGRA          Niagara
#> 12      <NA>            MAAC         72       SIENA            Siena
#> 13      <NA>            MAAC         80        IONA             Iona
#> 14      <NA>             WAC         88       TEXAS            Texas
#> 15      <NA>            MEAC         92        CORN          Cornell
#> 16      <NA>             NEC        111         HOU          Houston
#> 17      <NA>       Southland         80         ULL        Louisiana
#> 18      <NA>             WAC         69        UTEP             UTEP
#> 19      <NA>            UMAC        112      SDAKST South Dakota St.
#> 20      <NA>    NON-NCAA ORG         93      SO DAK     South Dakota
#> 21      <NA>             ACC         57         UVA         Virginia
#> 22      <NA>             AAC         63       BOISE        Boise St.
#> 23      <NA>         Big Ten         77      PURDUE           Purdue
#> 24      <NA>    NON-NCAA ORG         87      CSUBAK  CSU Bakersfield
#> 25      <NA>           SoCon         82        PORT         Portland
#> 26      <NA>   Summit League         76        SJSU     San Jose St.
#> 27      <NA>             WCC         86        UCSB UC Santa Barbara
#> 28      <NA>    NON-NCAA ORG         86      UC IRV        UC Irvine
#>           home_slug                                home_full home_conference
#> 1             nc-at      North Carolina A&T State University       Big South
#> 2    boston-college                           Boston College             ACC
#> 3            furman                        Furman University           SoCon
#> 4         south-fla              University of South Florida             AAC
#> 5          illinois  University of Illinois Urbana-Champaign         Big Ten
#> 6      presbyterian                     Presbyterian College       Big South
#> 7       st-johns-ny         St. John's University (New York)        Big East
#> 8      northeastern                  Northeastern University             CAA
#> 9        quinnipiac                    Quinnipiac University            MAAC
#> 10         canisius                         Canisius College            MAAC
#> 11          niagara                       Niagara University            MAAC
#> 12            siena                            Siena College            MAAC
#> 13             iona                             Iona College            MAAC
#> 14            texas            University of Texas at Austin          Big 12
#> 15          cornell                       Cornell University      Ivy League
#> 16          houston                    University of Houston             AAC
#> 17     la-lafayette     University of Louisiana at Lafayette        Sun Belt
#> 18             utep           University of Texas at El Paso           C-USA
#> 19  south-dakota-st            South Dakota State University   Summit League
#> 20     south-dakota               University of South Dakota   Summit League
#> 21         virginia                   University of Virginia             ACC
#> 22         boise-st                   Boise State University   Mountain West
#> 23           purdue                        Purdue University         Big Ten
#> 24      bakersfield California State University, Bakersfield        Big West
#> 25         portland                   University of Portland             WCC
#> 26      san-jose-st                San Jose State University   Mountain West
#> 27 uc-santa-barbara  University of California, Santa Barbara        Big West
#> 28        uc-irvine         University of California, Irvine        Big West
# }