ncaa_bb_player_box_score

Obtains basketball game player box scores from the NCAA API.

ncaa_bb_player_box_score(game_id)

Arguments

game_id

the game_id of a given NCAA BB game from the NCAA API

Value

a data frame with NCAA BB player box score data for the given game

Examples

# \donttest{
  ncaa_bb_player_box_score(5897924)
#>    team_loc    team_slug team_abbr team_short team_nick                    name
#> 1      away carver-bible    CARBIB     Carver                    Antwon Ferrell
#> 2      away carver-bible    CARBIB     Carver                        Glenn Sims
#> 3      away carver-bible    CARBIB     Carver                      Dyllon Scott
#> 4      away carver-bible    CARBIB     Carver                      Bryson Scott
#> 5      away carver-bible    CARBIB     Carver                   Stephon Auguste
#> 6      away carver-bible    CARBIB     Carver                    Keyshawn Coley
#> 7      away carver-bible    CARBIB     Carver           Sincere Jacobs-Hollomon
#> 8      away carver-bible    CARBIB     Carver                   Cerino Williams
#> 9      away carver-bible    CARBIB     Carver                    Cameron Knight
#> 10     away carver-bible    CARBIB     Carver              Brenden Middlebrooks
#> 11     away carver-bible    CARBIB     Carver                      Ka'jhun Gary
#> 12     away carver-bible    CARBIB     Carver                     Stephn Mayuen
#> 13     away carver-bible    CARBIB     Carver                       Chris Berry
#> 14     home        nc-at    NC A&T   N.C. A&T    Aggies         Kameron Langley
#> 15     home        nc-at    NC A&T   N.C. A&T    Aggies           Justin Brooks
#> 16     home        nc-at    NC A&T   N.C. A&T    Aggies              Trey Crews
#> 17     home        nc-at    NC A&T   N.C. A&T    Aggies          Ahmad Hamilton
#> 18     home        nc-at    NC A&T   N.C. A&T    Aggies               Kyle Duke
#> 19     home        nc-at    NC A&T   N.C. A&T    Aggies         Demetric Horton
#> 20     home        nc-at    NC A&T   N.C. A&T    Aggies           Marcus Watson
#> 21     home        nc-at    NC A&T   N.C. A&T    Aggies           Harry Morrice
#> 22     home        nc-at    NC A&T   N.C. A&T    Aggies         Jeremy Robinson
#> 23     home        nc-at    NC A&T   N.C. A&T    Aggies              Tyler Maye
#> 24     home        nc-at    NC A&T   N.C. A&T    Aggies         Milton Matthews
#> 25     home        nc-at    NC A&T   N.C. A&T    Aggies         Webster Filmore
#> 26     home        nc-at    NC A&T   N.C. A&T    Aggies            David Beatty
#> 27     home        nc-at    NC A&T   N.C. A&T    Aggies          Justin Whatley
#> 28     home        nc-at    NC A&T   N.C. A&T    Aggies          Chase McDuffie
#> 29     home        nc-at    NC A&T   N.C. A&T    Aggies          Jaylen Jackson
#> 30     home        nc-at    NC A&T   N.C. A&T    Aggies           Duncan Powell
#> 31     home        nc-at    NC A&T   N.C. A&T    Aggies        Jibril McCormick
#>    pos mp fgm fga     fgpct tpm tpa     tppct ftm fta     ftpct treb oreb dreb
#> 1    G 32   5  12 0.4166667   2   6 0.3333333   2   2 1.0000000    2    1    1
#> 2    G 28   1   4 0.2500000   0   1 0.0000000   0   1 0.0000000    3    1    2
#> 3    G 27   5  12 0.4166667   3   8 0.3750000   5   6 0.8333333    6    0    6
#> 4    G 23   2   7 0.2857143   0   4 0.0000000   2   3 0.6666667    2    0    2
#> 5    G 21   1   5 0.2000000   0   2 0.0000000   0   0       NaN    4    3    1
#> 6    G 20   0   3 0.0000000   0   1 0.0000000   2   2 1.0000000    1    0    1
#> 7    G 16   1   3 0.3333333   0   0       NaN   0   0       NaN    4    1    3
#> 8    G 14   2   6 0.3333333   0   1 0.0000000   0   0       NaN    4    1    3
#> 9    G 11   1   5 0.2000000   1   4 0.2500000   2   3 0.6666667    1    1    0
#> 10   G  0   0   0       NaN   0   0       NaN   0   0       NaN    0    0    0
#> 11   G  0   0   0       NaN   0   0       NaN   0   0       NaN    0    0    0
#> 12   G  0   0   0       NaN   0   0       NaN   0   0       NaN    0    0    0
#> 13   G  0   0   0       NaN   0   0       NaN   0   0       NaN    0    0    0
#> 14   G  7   1   2 0.5000000   0   0       NaN   1   2 0.5000000    0    0    0
#> 15   G  6   0   1 0.0000000   0   1 0.0000000   0   0       NaN    1    0    1
#> 16   G  6   0   0       NaN   0   0       NaN   4   4 1.0000000    0    0    0
#> 17   G  6   2   3 0.6666667   0   1 0.0000000   1   2 0.5000000    0    0    0
#> 18   G 27   3   8 0.3750000   0   1 0.0000000   3   4 0.7500000    5    2    3
#> 19   G 22   3   9 0.3333333   1   6 0.1666667   2   3 0.6666667    6    1    5
#> 20   G 20   2   5 0.4000000   0   1 0.0000000   0   0       NaN    4    2    2
#> 21   C 19   5   8 0.6250000   0   1 0.0000000   0   0       NaN    5    2    3
#> 22   F 18   4   7 0.5714286   0   2 0.0000000   1   3 0.3333333    6    3    3
#> 23   G 17   5   7 0.7142857   0   0       NaN   0   0       NaN    4    1    3
#> 24   G 17   4   9 0.4444444   2   6 0.3333333   0   0       NaN    6    0    6
#> 25   F 16   5   5 1.0000000   0   0       NaN   1   1 1.0000000    4    3    1
#> 26   G 15   5   8 0.6250000   3   5 0.6000000   0   0       NaN    2    0    2
#> 27   F  0   0   0       NaN   0   0       NaN   0   0       NaN    0    0    0
#> 28   G  0   0   0       NaN   0   0       NaN   0   0       NaN    0    0    0
#> 29   F  0   0   0       NaN   0   0       NaN   0   0       NaN    0    0    0
#> 30   F  0   0   0       NaN   0   0       NaN   0   0       NaN    0    0    0
#> 31   G  0   0   0       NaN   0   0       NaN   0   0       NaN    0    0    0
#>    ast pf stl to blk pts
#> 1    0  0   3  2   0  14
#> 2    1  2   3  5   0   2
#> 3    2  3   1  4   0  18
#> 4    0  1   1  2   0   6
#> 5    0  0   0  1   0   2
#> 6    0  2   0  0   1   2
#> 7    1  4   0  1   0   2
#> 8    0  5   0  1   0   4
#> 9    0  0   0  0   0   5
#> 10   0  0   0  0   0   0
#> 11   0  0   0  0   0   0
#> 12   0  0   0  0   0   0
#> 13   0  0   0  0   0   0
#> 14   1  0   1  2   0   3
#> 15   0  0   0  0   0   0
#> 16   0  1   0  0   0   4
#> 17   2  0   0  0   0   5
#> 18   5  2   1  1   0   9
#> 19   5  0   2  0   2   9
#> 20   1  1   0  2   0   4
#> 21   1  0   0  1   1  10
#> 22   1  5   0  1   0   9
#> 23   2  3   2  0   0  10
#> 24   2  1   2  2   1  10
#> 25   0  0   0  0   0  11
#> 26   0  2   1  1   0  13
#> 27   0  0   0  0   0   0
#> 28   0  0   0  0   0   0
#> 29   0  0   0  0   0   0
#> 30   0  0   0  0   0   0
#> 31   0  0   0  0   0   0
# }