ncaa_fb_games

Obtains a list of games for a given week and year of the season

ncaa_fb_games(week, year)

Arguments

week

the week number of the given season

year

the year for the season you wish to look at

Value

a data frame with all games for the given, week, year

Examples

# \donttest{
  ncaa_fb_games(10, 2021)
#>    game_game_id game_final_message game_bracket_round
#> 1        353837              FINAL                   
#> 2        353217              FINAL                   
#> 3        353570              FINAL                   
#> 4        353809              FINAL                   
#> 5        353872              FINAL                   
#> 6        353654              FINAL                   
#> 7        352572              FINAL                   
#> 8        353463              FINAL                   
#> 9        353987         FINAL (OT)                   
#> 10       353669              FINAL                   
#> 11       353245              FINAL                   
#> 12       352449              FINAL                   
#> 13       353043              FINAL                   
#> 14       353050              FINAL                   
#> 15       352825              FINAL                   
#> 16       353021              FINAL                   
#> 17       352734              FINAL                   
#> 18       352944              FINAL                   
#> 19       353169              FINAL                   
#> 20       352741              FINAL                   
#> 21       352464              FINAL                   
#> 22       353626              FINAL                   
#> 23       352609              FINAL                   
#> 24       353450              FINAL                   
#> 25       353895              FINAL                   
#> 26       352636              FINAL                   
#> 27       353308              FINAL                   
#> 28       353372              FINAL                   
#> 29       353830              FINAL                   
#> 30       353640              FINAL                   
#> 31       352429              FINAL                   
#> 32       353507              FINAL                   
#> 33       353486              FINAL                   
#> 34       353823              FINAL                   
#> 35       353015              FINAL                   
#> 36       353583              FINAL                   
#> 37       353140         FINAL (OT)                   
#> 38       353155              FINAL                   
#> 39       353181              FINAL                   
#> 40       352581              FINAL                   
#> 41       352805              FINAL                   
#> 42       352966              FINAL                   
#> 43       352672         FINAL (OT)                   
#> 44       353204              FINAL                   
#> 45       352797              FINAL                   
#> 46       353175              FINAL                   
#> 47       364350              FINAL                   
#> 48       352776              FINAL                   
#> 49       353036              FINAL                   
#> 50       353429              FINAL                   
#> 51       353767              FINAL                   
#> 52       352552              FINAL                   
#> 53       352790              FINAL                   
#> 54       352888              FINAL                   
#> 55       353401              FINAL                   
#> 56       352986              FINAL                   
#> 57       352444              FINAL                   
#> 58       353883              FINAL                   
#> 59       353703              FINAL                   
#> 60       352867              FINAL                   
#>                         game_title      game_url game_id game_network
#> 1                   Akron Ball St. /game/5851277 5851277             
#> 2                  Ohio Miami (OH) /game/5851278 5851278             
#> 3             Toledo Eastern Mich. /game/5851279 5851279             
#> 4      Western Mich. Central Mich. /game/5851280 5851280             
#> 5           Kent St. Northern Ill. /game/5851281 5851281             
#> 6            Louisiana Georgia St. /game/5851282 5851282             
#> 7     Boston College Virginia Tech /game/5851283 5851283             
#> 8                    Stanford Utah /game/5851284 5851284             
#> 9        Air Force Army West Point /game/5851323 5851323             
#> 10      North Carolina Wake Forest /game/5851296 5851296             
#> 11                Ole Miss Liberty /game/5851300 5851300             
#> 12              UAB Louisiana Tech /game/5851324 5851324             
#> 13         Miami (FL) Georgia Tech /game/5851311 5851311             
#> 14              Minnesota Illinois /game/5851290 5851290             
#> 15                Georgia Missouri /game/5851297 5851297             
#> 16                     Memphis SMU /game/5851306 5851306             
#> 17                 Duke Pittsburgh /game/5851308 5851308             
#> 18               Kansas Kansas St. /game/5851316 5851316             
#> 19               Nebraska Ohio St. /game/5851317 5851317             
#> 20            East Carolina Temple /game/5851287 5851287             
#> 21          Arkansas St. App State /game/5851304 5851304             
#> 22                   Texas St. ULM /game/5851321 5851321             
#> 23                   BYU Idaho St. /game/5851326 5851326             
#> 24      Southern Miss. North Texas /game/5851328 5851328             
#> 25      Massachusetts Rhode Island /game/5851285 5851285             
#> 26                Cincinnati Tulsa /game/5851315 5851315             
#> 27             Purdue Michigan St. /game/5851318 5851318             
#> 28               Rutgers Wisconsin /game/5851319 5851319             
#> 29            Wyoming Colorado St. /game/5851327 5851327             
#> 30                      UCF Tulane /game/5851293 5851293             
#> 31              Arizona California /game/5851314 5851314             
#> 32                      TCU Baylor /game/5851294 5851294             
#> 33                Texas A&M Auburn /game/5851301 5851301             
#> 34      West Virginia Oklahoma St. /game/5851312 5851312             
#> 35               Maryland Penn St. /game/5851320 5851320             
#> 36              Troy South Alabama /game/5851325 5851325             
#> 37                  Charlotte Rice /game/5851331 5851331             
#> 38                 Notre Dame Navy /game/5851332 5851332             
#> 39         New Mexico St. Utah St. /game/5851286 5851286             
#> 40        Arkansas Mississippi St. /game/5851289 5851289             
#> 41            Florida St. NC State /game/5851295 5851295             
#> 42              Kentucky Tennessee /game/5851288 5851288             
#> 43             Colorado Oregon St. /game/5851307 5851307             
#> 44               Northwestern Iowa /game/5851322 5851322             
#> 45            Fresno St. Boise St. /game/5851333 5851333             
#> 46                 New Mexico UNLV /game/5851334 5851334             
#> 47   Ga. Southern Coastal Carolina /game/5851335 5851335             
#> 48          Fla. Atlantic Marshall /game/5851336 5851336             
#> 49                Michigan Indiana /game/5851303 5851303             
#> 50              South Fla. Houston /game/5851305 5851305             
#> 51               Washington Oregon /game/5851309 5851309             
#> 52                     Alabama LSU /game/5851299 5851299             
#> 53                FIU Old Dominion /game/5851337 5851337             
#> 54                  Iowa St. Texas /game/5851292 5851292             
#> 55          South Carolina Florida /game/5851298 5851298             
#> 56              Louisville Clemson /game/5851302 5851302             
#> 57 Arizona St. Southern California /game/5851310 5851310             
#> 58             Nevada San Jose St. /game/5851313 5851313             
#> 59                       UTEP UTSA /game/5851329 5851329             
#> 60            Hawaii San Diego St. /game/5851291 5851291             
#>    game_live_video_enabled game_start_time game_start_time_epoch
#> 1                    FALSE      07:00PM ET            1635894000
#> 2                    FALSE      07:30PM ET            1635895800
#> 3                    FALSE      08:30PM ET            1635899400
#> 4                    FALSE      07:00PM ET            1635980400
#> 5                    FALSE      07:00PM ET            1635980400
#> 6                    FALSE      07:30PM ET            1636068600
#> 7                    FALSE      08:30PM ET            1636158600
#> 8                    FALSE      10:30PM ET            1636165800
#> 9                    FALSE      11:30AM ET            1636212600
#> 10                   FALSE      12:00PM ET            1636214400
#> 11                   FALSE      12:00PM ET            1636214400
#> 12                   FALSE      12:00PM ET            1636214400
#> 13                   FALSE      12:30PM ET            1636216200
#> 14                   FALSE      01:00PM ET            1636218000
#> 15                   FALSE      01:00PM ET            1636218000
#> 16                   FALSE      01:00PM ET            1636218000
#> 17                   FALSE      01:00PM ET            1636218000
#> 18                   FALSE      01:00PM ET            1636218000
#> 19                   FALSE      01:00PM ET            1636218000
#> 20                   FALSE      03:00PM ET            1636225200
#> 21                   FALSE      03:00PM ET            1636225200
#> 22                   FALSE      03:00PM ET            1636225200
#> 23                   FALSE      03:00PM ET            1636225200
#> 24                   FALSE      03:00PM ET            1636225200
#> 25                   FALSE      03:30PM ET            1636227000
#> 26                   FALSE      03:30PM ET            1636227000
#> 27                   FALSE      03:30PM ET            1636227000
#> 28                   FALSE      03:30PM ET            1636227000
#> 29                   FALSE      03:30PM ET            1636227000
#> 30                   FALSE      04:00PM ET            1636228800
#> 31                   FALSE      04:00PM ET            1636228800
#> 32                   FALSE      04:30PM ET            1636230600
#> 33                   FALSE      04:30PM ET            1636230600
#> 34                   FALSE      04:30PM ET            1636230600
#> 35                   FALSE      04:30PM ET            1636230600
#> 36                   FALSE      04:30PM ET            1636230600
#> 37                   FALSE      04:30PM ET            1636230600
#> 38                   FALSE      04:30PM ET            1636230600
#> 39                   FALSE      05:00PM ET            1636232400
#> 40                   FALSE      05:00PM ET            1636232400
#> 41                   FALSE      05:00PM ET            1636232400
#> 42                   FALSE      07:00PM ET            1636239600
#> 43                   FALSE      07:00PM ET            1636239600
#> 44                   FALSE      07:00PM ET            1636239600
#> 45                   FALSE      07:00PM ET            1636239600
#> 46                   FALSE      07:00PM ET            1636239600
#> 47                   FALSE      07:00PM ET            1636239600
#> 48                   FALSE      07:00PM ET            1636239600
#> 49                   FALSE      07:30PM ET            1636241400
#> 50                   FALSE      07:30PM ET            1636241400
#> 51                   FALSE      07:30PM ET            1636241400
#> 52                   FALSE      08:00PM ET            1636243200
#> 53                   FALSE      08:00PM ET            1636243200
#> 54                   FALSE      08:30PM ET            1636245000
#> 55                   FALSE      08:30PM ET            1636245000
#> 56                   FALSE      08:30PM ET            1636245000
#> 57                   FALSE      10:30PM ET            1636252200
#> 58                   FALSE      11:00PM ET            1636254000
#> 59                   FALSE      11:15PM ET            1636254900
#> 60                   FALSE      12:00AM ET            1636257600
#>    game_bracket_id game_game_state game_start_date game_current_period
#> 1                            final      11-02-2021               FINAL
#> 2                            final      11-02-2021               FINAL
#> 3                            final      11-02-2021               FINAL
#> 4                            final      11-03-2021               FINAL
#> 5                            final      11-03-2021               FINAL
#> 6                            final      11-04-2021               FINAL
#> 7                            final      11-05-2021               FINAL
#> 8                            final      11-05-2021               FINAL
#> 9                            final      11-06-2021          FINAL (OT)
#> 10                           final      11-06-2021               FINAL
#> 11                           final      11-06-2021               FINAL
#> 12                           final      11-06-2021               FINAL
#> 13                           final      11-06-2021               FINAL
#> 14                           final      11-06-2021               FINAL
#> 15                           final      11-06-2021               FINAL
#> 16                           final      11-06-2021               FINAL
#> 17                           final      11-06-2021               FINAL
#> 18                           final      11-06-2021               FINAL
#> 19                           final      11-06-2021               FINAL
#> 20                           final      11-06-2021               FINAL
#> 21                           final      11-06-2021               FINAL
#> 22                           final      11-06-2021               FINAL
#> 23                           final      11-06-2021               FINAL
#> 24                           final      11-06-2021               FINAL
#> 25                           final      11-06-2021               FINAL
#> 26                           final      11-06-2021               FINAL
#> 27                           final      11-06-2021               FINAL
#> 28                           final      11-06-2021               FINAL
#> 29                           final      11-06-2021               FINAL
#> 30                           final      11-06-2021               FINAL
#> 31                           final      11-06-2021               FINAL
#> 32                           final      11-06-2021               FINAL
#> 33                           final      11-06-2021               FINAL
#> 34                           final      11-06-2021               FINAL
#> 35                           final      11-06-2021               FINAL
#> 36                           final      11-06-2021               FINAL
#> 37                           final      11-06-2021          FINAL (OT)
#> 38                           final      11-06-2021               FINAL
#> 39                           final      11-06-2021               FINAL
#> 40                           final      11-06-2021               FINAL
#> 41                           final      11-06-2021               FINAL
#> 42                           final      11-06-2021               FINAL
#> 43                           final      11-06-2021          FINAL (OT)
#> 44                           final      11-06-2021               FINAL
#> 45                           final      11-06-2021               FINAL
#> 46                           final      11-06-2021               FINAL
#> 47                           final      11-06-2021               FINAL
#> 48                           final      11-06-2021               FINAL
#> 49                           final      11-06-2021               FINAL
#> 50                           final      11-06-2021               FINAL
#> 51                           final      11-06-2021               FINAL
#> 52                           final      11-06-2021               FINAL
#> 53                           final      11-06-2021               FINAL
#> 54                           final      11-06-2021               FINAL
#> 55                           final      11-06-2021               FINAL
#> 56                           final      11-06-2021               FINAL
#> 57                           final      11-06-2021               FINAL
#> 58                           final      11-06-2021               FINAL
#> 59                           final      11-06-2021               FINAL
#> 60                           final      11-07-2021               FINAL
#>    game_video_state game_bracket_region game_contest_clock game_away_score
#> 1                                                     0:00              31
#> 2                                                    12:18              33
#> 3                                                     0:00              52
#> 4                                                     0:00              42
#> 5                                                     0:00              47
#> 6                                                     0:00              17
#> 7                                                     0:00               3
#> 8                                                     0:00              52
#> 9                                                                       21
#> 10                                                    0:00              55
#> 11                                                    0:00              14
#> 12                                                    0:00              38
#> 13                                                    0:00              30
#> 14                                                    0:00              14
#> 15                                                    0:00               6
#> 16                                                    0:00              25
#> 17                                                    0:00              54
#> 18                                                    0:00              35
#> 19                                                    0:00              26
#> 20                                                    0:00               3
#> 21                                                    0:00              48
#> 22                                                    0:00              19
#> 23                                                    0:00              14
#> 24                                                    0:00              38
#> 25                                                    0:00              35
#> 26                                                    0:00              20
#> 27                                                    0:00              29
#> 28                                                    0:00              52
#> 29                                                    0:00              17
#> 30                                                    0:00              10
#> 31                                                    0:00               3
#> 32                                                    0:00              28
#> 33                                                    0:00               3
#> 34                                                    0:00              24
#> 35                                                    0:00              31
#> 36                                                    0:00              24
#> 37                                                                      24
#> 38                                                    0:00               6
#> 39                                                    0:00              35
#> 40                                                    0:00              28
#> 41                                                    0:00              28
#> 42                                                    0:00              45
#> 43                                                                      34
#> 44                                                    0:00              17
#> 45                                                    0:00              40
#> 46                                                    0:00              31
#> 47                                                    0:00              28
#> 48                                                    0:00              28
#> 49                                                    0:00               7
#> 50                                                    0:00              60
#> 51                                                    0:00              26
#> 52                                                    0:00              14
#> 53                                                    0:00              47
#> 54                                                    0:00               7
#> 55                                                    0:00              17
#> 56                                                    0:00              30
#> 57                                                    0:00              16
#> 58                                                    0:00              20
#> 59                                                    0:00              44
#> 60                                                    0:00              17
#>    game_away_winner game_away_seed game_away_description game_away_rank
#> 1              TRUE                                (5-4)               
#> 2             FALSE                                (4-5)               
#> 3              TRUE                                (6-3)               
#> 4              TRUE                                (5-4)               
#> 5             FALSE                                (6-3)               
#> 6             FALSE                                (4-5)               
#> 7             FALSE                                (4-5)               
#> 8              TRUE                                (6-3)               
#> 9              TRUE                                (5-3)               
#> 10            FALSE                                (8-1)              9
#> 11            FALSE                                (7-3)               
#> 12            FALSE                                (2-7)               
#> 13            FALSE                                (3-6)               
#> 14             TRUE                                (4-6)               
#> 15            FALSE                                (4-5)               
#> 16            FALSE                                (7-2)               
#> 17             TRUE                                (7-2)             25
#> 18             TRUE                                (6-3)               
#> 19             TRUE                                (8-1)              5
#> 20            FALSE                                (3-6)               
#> 21             TRUE                                (7-2)               
#> 22            FALSE                                (4-5)               
#> 23            FALSE                                (1-8)               
#> 24             TRUE                                (3-6)               
#> 25             TRUE                                (6-3)               
#> 26            FALSE                                (3-6)               
#> 27            FALSE                                (8-1)              3
#> 28             TRUE                                (6-3)             21
#> 29            FALSE                                (3-6)               
#> 30            FALSE                                (1-8)               
#> 31            FALSE                                (3-6)               
#> 32            FALSE                                (7-2)             12
#> 33            FALSE                                (6-3)             13
#> 34             TRUE                                (8-1)             11
#> 35             TRUE                                (6-3)               
#> 36            FALSE                                (5-4)               
#> 37            FALSE                                (3-6)               
#> 38            FALSE                                (2-7)               
#> 39             TRUE                                (7-2)               
#> 40            FALSE                                (5-4)             17
#> 41             TRUE                                (7-2)             19
#> 42             TRUE                                (5-4)               
#> 43            FALSE                                (5-4)               
#> 44             TRUE                                (7-2)             22
#> 45             TRUE                                (5-4)               
#> 46             TRUE                                (1-8)               
#> 47             TRUE                                (8-1)               
#> 48             TRUE                                (6-3)               
#> 49            FALSE                                (2-7)               
#> 50             TRUE                                (8-1)               
#> 51             TRUE                                (8-1)              4
#> 52            FALSE                                (4-5)               
#> 53             TRUE                                (3-6)               
#> 54            FALSE                                (4-5)               
#> 55            FALSE                                (4-5)               
#> 56             TRUE                                (6-3)               
#> 57            FALSE                                (4-5)               
#> 58            FALSE                                (5-5)               
#> 59             TRUE                                (9-0)               
#> 60             TRUE                                (8-1)             24
#>    game_away_names_char6 game_away_names_short game_away_names_seo
#> 1                 BALLST              Ball St.             ball-st
#> 2                 MIA OH            Miami (OH)            miami-oh
#> 3                 E MICH         Eastern Mich.        eastern-mich
#> 4                 C MICH         Central Mich.        central-mich
#> 5                    NIU         Northern Ill.        northern-ill
#> 6                  GA ST           Georgia St.          georgia-st
#> 7                     VT         Virginia Tech       virginia-tech
#> 8                   UTAH                  Utah                utah
#> 9                   ARMY       Army West Point                army
#> 10                  WAKE           Wake Forest         wake-forest
#> 11                LIBRTY               Liberty             liberty
#> 12                LATECH        Louisiana Tech      louisiana-tech
#> 13                GATECH          Georgia Tech        georgia-tech
#> 14                   ILL              Illinois            illinois
#> 15                MIZZOU              Missouri            missouri
#> 16                   SMU                   SMU                 smu
#> 17                  PITT            Pittsburgh          pittsburgh
#> 18                KAN ST            Kansas St.           kansas-st
#> 19                OHIOST              Ohio St.             ohio-st
#> 20                TEMPLE                Temple              temple
#> 21                APP ST             App State      appalachian-st
#> 22                LA MON                   ULM           la-monroe
#> 23                 ID ST             Idaho St.            idaho-st
#> 24                NO TEX           North Texas         north-texas
#> 25                   URI          Rhode Island        rhode-island
#> 26                 TULSA                 Tulsa               tulsa
#> 27                MICHST          Michigan St.         michigan-st
#> 28                  WISC             Wisconsin           wisconsin
#> 29                 CO ST          Colorado St.         colorado-st
#> 30                TULANE                Tulane              tulane
#> 31                   CAL            California          california
#> 32                BAYLOR                Baylor              baylor
#> 33                AUBURN                Auburn              auburn
#> 34                OKLAST          Oklahoma St.         oklahoma-st
#> 35                PENNST              Penn St.             penn-st
#> 36                 S ALA         South Alabama           south-ala
#> 37                  RICE                  Rice                rice
#> 38                  NAVY                  Navy                navy
#> 39                 UT ST              Utah St.             utah-st
#> 40                MISSST       Mississippi St.      mississippi-st
#> 41                 NC ST              NC State   north-carolina-st
#> 42                  TENN             Tennessee           tennessee
#> 43                ORE ST            Oregon St.           oregon-st
#> 44                  IOWA                  Iowa                iowa
#> 45                 BOISE             Boise St.            boise-st
#> 46                  UNLV                  UNLV                unlv
#> 47                CO CAR      Coastal Carolina        coastal-caro
#> 48                 MARSH              Marshall            marshall
#> 49                   IND               Indiana             indiana
#> 50                   HOU               Houston             houston
#> 51                 OREGN                Oregon              oregon
#> 52                   LSU                   LSU                 lsu
#> 53                   ODU          Old Dominion        old-dominion
#> 54                 TEXAS                 Texas               texas
#> 55                   FLA               Florida             florida
#> 56                  CLEM               Clemson             clemson
#> 57                   USC   Southern California southern-california
#> 58                  SJSU          San Jose St.         san-jose-st
#> 59                  UTSA                  UTSA                utsa
#> 60                  SDSU         San Diego St.        san-diego-st
#>                                   game_away_names_full game_home_score
#> 1                                Ball State University              25
#> 2                              Miami University (Ohio)              35
#> 3                          Eastern Michigan University              49
#> 4                          Central Michigan University              30
#> 5                         Northern Illinois University              52
#> 6                             Georgia State University              21
#> 7  Virginia Polytechnic Institute and State University              17
#> 8                                   University of Utah               7
#> 9                                U.S. Military Academy              14
#> 10                              Wake Forest University              58
#> 11                                  Liberty University              27
#> 12                           Louisiana Tech University              52
#> 13                     Georgia Institute of Technology              33
#> 14             University of Illinois Urbana-Champaign               6
#> 15                    University of Missouri, Columbia              43
#> 16                       Southern Methodist University              28
#> 17                            University of Pittsburgh              29
#> 18                             Kansas State University              10
#> 19                           The Ohio State University              17
#> 20                                   Temple University              45
#> 21                        Appalachian State University              14
#> 22                      University of Louisiana Monroe              27
#> 23                              Idaho State University              59
#> 24                           University of North Texas              14
#> 25                          University of Rhode Island              22
#> 26                             The University of Tulsa              28
#> 27                           Michigan State University              40
#> 28                     University of Wisconsin-Madison               3
#> 29                           Colorado State University              31
#> 30                                   Tulane University              14
#> 31                  University of California, Berkeley              10
#> 32                                   Baylor University              30
#> 33                                   Auburn University              20
#> 34                           Oklahoma State University               3
#> 35                       Pennsylvania State University              14
#> 36                         University of South Alabama              31
#> 37                                     Rice University              31
#> 38                                  U.S. Naval Academy              34
#> 39                               Utah State University              13
#> 40                        Mississippi State University              31
#> 41                     North Carolina State University              14
#> 42                  University of Tennessee, Knoxville              42
#> 43                             Oregon State University              37
#> 44                                  University of Iowa              12
#> 45                              Boise State University              14
#> 46                     University of Nevada, Las Vegas              17
#> 47                         Coastal Carolina University               8
#> 48                                 Marshall University              13
#> 49                     Indiana University, Bloomington              29
#> 50                               University of Houston              42
#> 51                                University of Oregon              16
#> 52                          Louisiana State University              20
#> 53                             Old Dominion University              24
#> 54                       University of Texas at Austin              30
#> 55                               University of Florida              40
#> 56                                  Clemson University              24
#> 57                   University of Southern California              31
#> 58                           San Jose State University              51
#> 59                  University of Texas at San Antonio              23
#> 60                          San Diego State University              10
#>    game_home_winner game_home_seed game_home_description game_home_rank
#> 1             FALSE                                (2-7)               
#> 2              TRUE                                (2-7)               
#> 3             FALSE                                (4-5)               
#> 4             FALSE                                (5-4)               
#> 5              TRUE                                (5-4)               
#> 6              TRUE                                (8-1)               
#> 7              TRUE                                (5-4)               
#> 8             FALSE                                (3-6)               
#> 9             FALSE                                (6-3)               
#> 10             TRUE                                (5-4)               
#> 11             TRUE                                (7-2)             16
#> 12             TRUE                                (6-3)               
#> 13             TRUE                                (5-4)               
#> 14            FALSE                                (6-3)             20
#> 15             TRUE                                (9-0)              1
#> 16             TRUE                                (5-4)               
#> 17            FALSE                                (3-6)               
#> 18            FALSE                                (1-8)               
#> 19            FALSE                                (3-7)               
#> 20             TRUE                                (5-4)               
#> 21            FALSE                                (1-8)               
#> 22             TRUE                                (3-6)               
#> 23             TRUE                                (8-2)             15
#> 24            FALSE                                (1-8)               
#> 25            FALSE                                (1-8)               
#> 26             TRUE                                (9-0)              6
#> 27             TRUE                                (6-3)               
#> 28            FALSE                                (4-5)               
#> 29             TRUE                                (5-4)               
#> 30             TRUE                                (6-3)               
#> 31             TRUE                                (1-8)               
#> 32             TRUE                                (4-5)               
#> 33             TRUE                                (7-2)             14
#> 34            FALSE                                (4-5)               
#> 35            FALSE                                (5-4)               
#> 36             TRUE                                (5-4)               
#> 37             TRUE                                (5-4)               
#> 38             TRUE                                (8-1)             10
#> 39            FALSE                                (1-8)               
#> 40             TRUE                                (6-3)               
#> 41            FALSE                                (3-6)               
#> 42            FALSE                                (6-3)             18
#> 43             TRUE                                (3-6)               
#> 44            FALSE                                (3-6)               
#> 45            FALSE                                (7-3)             23
#> 46            FALSE                                (3-6)               
#> 47            FALSE                                (2-7)               
#> 48            FALSE                                (5-4)               
#> 49             TRUE                                (8-1)              7
#> 50            FALSE                                (2-7)               
#> 51            FALSE                                (4-5)               
#> 52             TRUE                                (8-1)              2
#> 53            FALSE                                (1-8)               
#> 54             TRUE                                (6-3)               
#> 55             TRUE                                (5-4)               
#> 56            FALSE                                (4-5)               
#> 57             TRUE                                (6-3)               
#> 58             TRUE                                (7-2)               
#> 59            FALSE                                (6-3)               
#> 60            FALSE                                (4-6)               
#>    game_home_names_char6 game_home_names_short game_home_names_seo
#> 1                  AKRON                 Akron               akron
#> 2                   OHIO                  Ohio                ohio
#> 3                 TOLEDO                Toledo              toledo
#> 4                 W MICH         Western Mich.        western-mich
#> 5                   KENT              Kent St.             kent-st
#> 6                    ULL             Louisiana        la-lafayette
#> 7                     BC        Boston College      boston-college
#> 8                   STAN              Stanford            stanford
#> 9                 AIRFOR             Air Force           air-force
#> 10                   UNC        North Carolina      north-carolina
#> 11                  MISS              Ole Miss            ole-miss
#> 12                   UAB                   UAB                 uab
#> 13                 MIAMI            Miami (FL)            miami-fl
#> 14                  MINN             Minnesota           minnesota
#> 15                   UGA               Georgia             georgia
#> 16                   MEM               Memphis             memphis
#> 17                  DUKE                  Duke                duke
#> 18                KANSAS                Kansas              kansas
#> 19                   NEB              Nebraska            nebraska
#> 20                 E CAR         East Carolina       east-carolina
#> 21                ARK ST          Arkansas St.         arkansas-st
#> 22                 TX ST             Texas St.            texas-st
#> 23                   BYU                   BYU                 byu
#> 24                S MISS        Southern Miss.       southern-miss
#> 25                 UMASS         Massachusetts       massachusetts
#> 26                 CINCY            Cincinnati          cincinnati
#> 27                PURDUE                Purdue              purdue
#> 28                RUTGER               Rutgers             rutgers
#> 29                   WYO               Wyoming             wyoming
#> 30                   UCF                   UCF                 ucf
#> 31                  ARIZ               Arizona             arizona
#> 32                   TCU                   TCU                 tcu
#> 33                TX A&M             Texas A&M            texas-am
#> 34                   WVU         West Virginia       west-virginia
#> 35                    MD              Maryland            maryland
#> 36                  TROY                  Troy                troy
#> 37                  CHAR             Charlotte           charlotte
#> 38                N DAME            Notre Dame          notre-dame
#> 39                 NM ST        New Mexico St.       new-mexico-st
#> 40                   ARK              Arkansas            arkansas
#> 41                   FSU           Florida St.          florida-st
#> 42                    UK              Kentucky            kentucky
#> 43                  COLO              Colorado            colorado
#> 44                NWSTRN          Northwestern        northwestern
#> 45                FRESNO            Fresno St.           fresno-st
#> 46                 N MEX            New Mexico          new-mexico
#> 47                GA SOU          Ga. Southern         ga-southern
#> 48                   FAU         Fla. Atlantic        fla-atlantic
#> 49                  MICH              Michigan            michigan
#> 50                   USF            South Fla.           south-fla
#> 51                  WASH            Washington          washington
#> 52                   ALA               Alabama             alabama
#> 53                   FIU                   FIU                 fiu
#> 54                IOWAST              Iowa St.             iowa-st
#> 55                 S CAR        South Carolina      south-carolina
#> 56                 LOUIS            Louisville          louisville
#> 57                 AZ ST           Arizona St.          arizona-st
#> 58                NEVADA                Nevada              nevada
#> 59                  UTEP                  UTEP                utep
#> 60                HAWAII                Hawaii              hawaii
#>                                          game_home_names_full away_conference
#> 1                                         University of Akron             MAC
#> 2                                             Ohio University             MAC
#> 3                                        University of Toledo             MAC
#> 4                                 Western Michigan University             MAC
#> 5                                       Kent State University             MAC
#> 6                        University of Louisiana at Lafayette        Sun Belt
#> 7                                              Boston College             ACC
#> 8                                         Stanford University          Pac-12
#> 9                                      U.S. Air Force Academy FBS Independent
#> 10                  University of North Carolina, Chapel Hill             ACC
#> 11                                  University of Mississippi FBS Independent
#> 12                        University of Alabama at Birmingham           C-USA
#> 13                              University of Miami (Florida)             ACC
#> 14                       University of Minnesota, Twin Cities         Big Ten
#> 15                                      University of Georgia             SEC
#> 16                                      University of Memphis             AAC
#> 17                                            Duke University             ACC
#> 18                                       University of Kansas          Big 12
#> 19                            University of Nebraska, Lincoln         Big Ten
#> 20                                   East Carolina University             AAC
#> 21                                  Arkansas State University        Sun Belt
#> 22                                     Texas State University        Sun Belt
#> 23                                   Brigham Young University         Big Sky
#> 24                     The University of Southern Mississippi           C-USA
#> 25                       University of Massachusetts, Amherst             CAA
#> 26                                   University of Cincinnati             AAC
#> 27                                          Purdue University         Big Ten
#> 28 Rutgers, The State University of New Jersey, New Brunswick         Big Ten
#> 29                                      University of Wyoming   Mountain West
#> 30                              University of Central Florida             AAC
#> 31                                      University of Arizona          Pac-12
#> 32                                 Texas Christian University          Big 12
#> 33                      Texas A&M University, College Station             SEC
#> 34                                   West Virginia University          Big 12
#> 35                       University of Maryland, College Park         Big Ten
#> 36                                            Troy University        Sun Belt
#> 37              The University of North Carolina at Charlotte           C-USA
#> 38                                   University of Notre Dame             AAC
#> 39                                New Mexico State University   Mountain West
#> 40                       University of Arkansas, Fayetteville             SEC
#> 41                                   Florida State University             ACC
#> 42                                     University of Kentucky             SEC
#> 43                            University of Colorado, Boulder          Pac-12
#> 44                                    Northwestern University         Big Ten
#> 45                        California State University, Fresno   Mountain West
#> 46                                   University of New Mexico   Mountain West
#> 47                                Georgia Southern University        Sun Belt
#> 48                                Florida Atlantic University           C-USA
#> 49                                     University of Michigan         Big Ten
#> 50                                University of South Florida             AAC
#> 51                                   University of Washington          Pac-12
#> 52                                      University of Alabama             SEC
#> 53                           Florida International University           C-USA
#> 54                                      Iowa State University          Big 12
#> 55                     University of South Carolina, Columbia             SEC
#> 56                                   University of Louisville             ACC
#> 57                                   Arizona State University          Pac-12
#> 58                                 University of Nevada, Reno   Mountain West
#> 59                             University of Texas at El Paso           C-USA
#> 60                                University of Hawaii, Manoa   Mountain West
#>    home_conference
#> 1              MAC
#> 2              MAC
#> 3              MAC
#> 4              MAC
#> 5              MAC
#> 6         Sun Belt
#> 7              ACC
#> 8           Pac-12
#> 9    Mountain West
#> 10             ACC
#> 11             SEC
#> 12           C-USA
#> 13             ACC
#> 14         Big Ten
#> 15             SEC
#> 16             AAC
#> 17             ACC
#> 18          Big 12
#> 19         Big Ten
#> 20             AAC
#> 21        Sun Belt
#> 22        Sun Belt
#> 23 FBS Independent
#> 24           C-USA
#> 25 FBS Independent
#> 26             AAC
#> 27         Big Ten
#> 28         Big Ten
#> 29   Mountain West
#> 30             AAC
#> 31          Pac-12
#> 32          Big 12
#> 33             SEC
#> 34          Big 12
#> 35         Big Ten
#> 36        Sun Belt
#> 37           C-USA
#> 38 FBS Independent
#> 39 FBS Independent
#> 40             SEC
#> 41             ACC
#> 42             SEC
#> 43          Pac-12
#> 44         Big Ten
#> 45   Mountain West
#> 46   Mountain West
#> 47        Sun Belt
#> 48           C-USA
#> 49         Big Ten
#> 50             AAC
#> 51          Pac-12
#> 52             SEC
#> 53           C-USA
#> 54          Big 12
#> 55             SEC
#> 56             ACC
#> 57          Pac-12
#> 58   Mountain West
#> 59           C-USA
#> 60   Mountain West
# }