ncaa_rosters

a generalized function for obtaining NCAA rosters for individual teams

ncaa_rosters(team_id)

Arguments

team_id

a unique team identifier for a given NCAA team from any sport

Value

a dataframe of roster data with the team, team_id, player, player_id, pos, year, games_played, and games_started values for each player on a given team. Where available, the height of players is returned, in inches.

Examples

# \donttest{
ncaa_rosters('505669')
#> # A tibble: 17 × 10
#>    team        team_id player   player_id pos   year  height jersey games_played
#>    <chr>       <chr>   <chr>        <int> <chr> <chr>  <int>  <int>        <int>
#>  1 Georgia Te… 505669  Alvarad…   1956747 G     Sr        72     10           26
#>  2 Georgia Te… 505669  Boyd, C…   2266724 G     So        73     13            2
#>  3 Georgia Te… 505669  Broadwa…   2280760 G     So        75     34            2
#>  4 Georgia Te… 505669  Devoe, …   2081315 G     Jr        77      0           26
#>  5 Georgia Te… 505669  Didenko…   2266725 F     Jr        81     14            5
#>  6 Georgia Te… 505669  Gigiber…   2447525 C     Fr        85      2            9
#>  7 Georgia Te… 505669  Howard,…   2272411 C     So        82     24           19
#>  8 Georgia Te… 505669  James, …   2280742 F     So        78     35            4
#>  9 Georgia Te… 505669  Maxwell…   2447453 G     Fr        74     11            1
#> 10 Georgia Te… 505669  Medlock…   1845892 G     Sr        77     45            2
#> 11 Georgia Te… 505669  Meka, J…   2447526 C     Fr        80     23            1
#> 12 Georgia Te… 505669  Moore, …   2081283 F     Jr        79     12           25
#> 13 Georgia Te… 505669  Parham,…   1965493 G     Sr        71      3           25
#> 14 Georgia Te… 505669  Rice, M…   1956752 G     Sr        72     55            1
#> 15 Georgia Te… 505669  Sturdiv…   2272128 G     So        74      1           26
#> 16 Georgia Te… 505669  Usher, …   1962448 F     Sr        79      4           26
#> 17 Georgia Te… 505669  Wright,…   1956749 F     Sr        81      5           25
#> # … with 1 more variable: games_started <int>
# }